ai-services/local-llm/requirements.txt
requests==2.31.0
psutil==5.9.6
loguru==0.7.2
python-dotenv==1.0.0
aiohttp==3.9.1
